move handleCatalysts to standalone function
